Attempt to remove unneeded packages only if the removal list is non-empty.

+++ ncm-spma/src/main/perl/spma/yumng.pm
@@ -688,12 +688,14 @@ sub Configure
         my $installed_names = Set::Scalar->new( split ( /\n/, $cmd_out ) );
         my $packages_to_remove = $installed_names - $to_install_names;
         $packages_to_remove->delete('gpg-pubkey');
-        ( $cmd_exit, $cmd_out, $cmd_err ) = $self->execute_command([ "yum remove -y " . YUM_PLUGIN_OPTS . " " . join (" ", sort @$packages_to_remove) ], 'attempting to remove: '.join (" ", sort @$packages_to_remove));
-        $installed = $self->get_installed_rpms();
-        return 1 if !defined($installed);
-        my $removed = $preinstalled - $installed;
-        if ( !$removed->is_empty ) {
-            $self->info("removed " . $removed->size . " unneeded package(s): " . $removed );
+        if ( !$packages_to_remove->is_empty ) {
+            ( $cmd_exit, $cmd_out, $cmd_err ) = $self->execute_command([ "yum remove -y " . YUM_PLUGIN_OPTS . " " . join (" ", sort @$packages_to_remove) ], 'attempting to remove: '.join (" ", sort @$packages_to_remove));
+            $installed = $self->get_installed_rpms();
+            return 1 if !defined($installed);
+            my $removed = $preinstalled - $installed;
+            if ( !$removed->is_empty ) {
+                $self->info("removed " . $removed->size . " unneeded package(s): " . $removed );
+            }
         }
     }
